requirements.txt
kafka-python
pyyaml
python-dotenv